refactor to decoders [skip ci]
