+++ crates/polars-parquet/src/arrow/read/deserialize/binary/mod.rs
@@ -2,8 +2,10 @@ mod basic;
 mod dictionary;
 mod nested;
 mod utils;
+mod decoders;
 
 pub use basic::Iter;
 pub use dictionary::{DictIter, NestedDictIter};
 pub use nested::NestedIter;
 pub(super) use utils::BinaryIter;
+
